pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bomer"
version = "0.1.0"
description = "Bomer – deterministic BoM analysis and optimization engine"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Emre Can Yesilyurt" }
]
dependencies = [
  "pandas>=1.5",
  "pyyaml>=6.0",
]

[project.scripts]
bomer = "bomer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
